# specs/002-microservices-infra/contracts/story-service-openapi.yml
openapi: 3.0.3
info:
  title: Story Service API
  version: v2.0.0
  description: |
    Story Service provides CRUD operations for stories and integration with photo personalization.
    Supports User Story 1 (View and Experience Personalized Story) from StoryMe app specification.

servers:
  - url: http://story-service:8000
    description: Internal service endpoint (accessed via API Gateway)
  - url: https://api.storyme.example.com/stories
    description: Public API Gateway endpoint

paths:
  /health:
    get:
      summary: Health check endpoint
      description: Returns 200 if service is healthy, 503 if unhealthy (FR-034)
      operationId: healthCheck
      tags:
        - Health
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [healthy]
                  version:
                    type: string
                    example: "v2.0.0"
                  uptime_seconds:
                    type: integer
                    example: 3600
        '503':
          description: Service is unhealthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [unhealthy]
                  error:
                    type: string

  /stories:
    get:
      summary: List stories
      description: Get paginated list of stories, filtered by subscription tier if user authenticated
      operationId: listStories
      tags:
        - Stories
      parameters:
        - name: user_id
          in: query
          description: User ID to filter available stories by subscription
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: List of stories
          content:
            application/json:
              schema:
                type: object
                properties:
                  stories:
                    type: array
                    items:
                      $ref: '#/components/schemas/Story'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      summary: Create new story
      description: Create a new story (admin only)
      operationId: createStory
      tags:
        - Stories
      security:
        - jwtAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, content, illustration_urls]
              properties:
                title:
                  type: string
                  minLength: 1
                  maxLength: 200
                content:
                  type: string
                  description: Story narrative text
                illustration_urls:
                  type: array
                  items:
                    type: string
                    format: uri
                narration_url:
                  type: string
                  format: uri
                age_range:
                  type: string
                  enum: [3-5, 6-8, 9-10]
                tier:
                  type: string
                  enum: [free, subscriber]
                  default: subscriber
      responses:
        '201':
          description: Story created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Story'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

  /stories/{story_id}:
    get:
      summary: Get story by ID
      description: Fetch a single story with personalization if available
      operationId: getStory
      tags:
        - Stories
      parameters:
        - name: story_id
          in: path
          required: true
          schema:
            type: string
        - name: user_id
          in: query
          description: User ID to apply personalization
          schema:
            type: string
      responses:
        '200':
          description: Story details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoryWithPersonalization'
        '404':
          description: Story not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Story not found"
        '500':
          $ref: '#/components/responses/InternalError'

  /stories/{story_id}/personalize:
    post:
      summary: Apply personalization to story
      description: Trigger face-swap personalization for a story using user's uploaded photo
      operationId: personalizeStory
      tags:
        - Personalization
      security:
        - jwtAuth: []
      parameters:
        - name: story_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id, photo_id]
              properties:
                user_id:
                  type: string
                photo_id:
                  type: string
                  description: Photo ID from photo-service
      responses:
        '202':
          description: Personalization request accepted (async processing)
          content:
            application/json:
              schema:
                type: object
                properties:
                  personalization_id:
                    type: string
                  status:
                    type: string
                    enum: [pending]
                  estimated_completion_seconds:
                    type: integer
                    example: 30
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Story or photo not found
        '500':
          $ref: '#/components/responses/InternalError'

components:
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Story:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        content:
          type: string
        illustration_urls:
          type: array
          items:
            type: string
            format: uri
        narration_url:
          type: string
          format: uri
        age_range:
          type: string
          enum: [3-5, 6-8, 9-10]
        tier:
          type: string
          enum: [free, subscriber]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    StoryWithPersonalization:
      allOf:
        - $ref: '#/components/schemas/Story'
        - type: object
          properties:
            personalization:
              type: object
              properties:
                status:
                  type: string
                  enum: [none, pending, completed, failed]
                personalized_illustrations:
                  type: array
                  items:
                    type: string
                    format: uri
                  description: URLs to illustrations with face-swap applied

    Pagination:
      type: object
      properties:
        current_page:
          type: integer
        total_pages:
          type: integer
        total_items:
          type: integer
        items_per_page:
          type: integer

  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Validation failed"
              details:
                type: array
                items:
                  type: object
                  properties:
                    field:
                      type: string
                    message:
                      type: string

    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Unauthorized"
              message:
                type: string
                example: "Valid authentication required"

    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Internal server error"
              trace_id:
                type: string
                description: Trace ID for debugging (FR-010)
                example: "550e8400-e29b-41d4-a716-446655440000"

# Service Dependencies (for contract testing - FR-030)
x-service-dependencies:
  payment-service:
    version: ">=v1.0.0, <v2.0.0"
    endpoints:
      - path: /subscriptions/check
        method: GET
        description: "Check if user is subscriber to filter story access"

  photo-service:
    version: ">=v1.0.0, <v2.0.0"
    endpoints:
      - path: /photos/{id}/status
        method: GET
        description: "Check photo processing status"
      - path: /personalize
        method: POST
        description: "Request face-swap personalization"